torch/requirements.txt
torch==2.2.2+cu121
numpy==1.26.4
matplotlib==3.9.0
opencv-python==4.10.0.84
omegaconf==2.3.0
hydra-core==1.3.2
einops==0.8.0
tensorboardX==2.6.2.2
torchvision==0.17.2+cu121
timm==1.0.14